+++ tket/include/tket/Circuit/Circuit.hpp
@@ -836,9 +836,9 @@ class Circuit {
   Vertex add_op(
       OpType type, const std::vector<Expr> &params, const std::vector<ID> &args,
       std::optional<std::string> opgroup = std::nullopt) {
-    if (is_metaop_type(type)) {
+    if (is_metaop_type(type) || is_barrier_type(type)) {
       throw CircuitInvalidity(
-          "Cannot add metaop. Please use `add_barrier` to add a "
+          "Cannot add metaop or barrier. Please use `add_barrier` to add a "
           "barrier.");
     }
     return add_op(get_op_ptr(type, params, args.size()), args, opgroup);
@@ -904,6 +904,11 @@ class Circuit {
     if (is_metaop_type(type)) {
       throw CircuitInvalidity("Cannot add a conditional metaop.");
     }
+    if (is_barrier_type(type)) {
+      throw CircuitInvalidity(
+          "Please use 'add_conditional_barrier' to add a conditional barrier "
+          "gate.");
+    }
     Op_ptr cond = std::make_shared<Conditional>(
         get_op_ptr(type, params, (unsigned)args.size()), (unsigned)bits.size(),
         value);
@@ -917,6 +922,18 @@ class Circuit {
       const std::vector<unsigned> &bits = {}, const std::string &_data = "");
 
   Vertex add_barrier(const unit_vector_t &args, const std::string &_data = "");
+
+  Vertex add_conditional_barrier(
+      const std::vector<unsigned> &barrier_qubits,
+      const std::vector<unsigned> &barrier_bits,
+      const std::vector<unsigned> &condition_bits, unsigned value,
+      const std::string &_data,
+      std::optional<std::string> opgroup = std::nullopt);
+
+  Vertex add_conditional_barrier(
+      const unit_vector_t &barrier_args, const bit_vector_t &condition_bits,
+      unsigned value, const std::string &_data,
+      std::optional<std::string> opgroup = std::nullopt);
 
   /**
    * Add a postfix to a classical register name if the register exists